Rename file for savedTiles
